+++ CHANGELOG.md
@@ -27,6 +27,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+- Prepare PT for es validation (address parser, synonyms, zip exclusion) [#46](https://github.com/Shopify/atlas_engine/pull/46)
 - Improve ingestion and validation for Slovenia [#76](https://github.com/Shopify/atlas_engine/pull/76)
 - Enable Sorbet/ForbidTStruct cop [#78] (https://github.com/Shopify/atlas_engine/pull/78)
 - Fix lookup of unmatched_components_suggestion_threshold [#73](https://github.com/Shopify/atlas_engine/pull/73)